services/ocr-service/requirements.txt
# --- requirements.txt ---
fastapi==0.111.0
uvicorn==0.30.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
numpy==1.26.4
aiofiles==23.2.0
rq
redis
